RAG_Chatbot/requirements.txt
langchain==0.1.17
langchain-groq==0.1.3
llama-parse==0.1.3
qdrant-client==1.9.1
unstructured[md]==0.13.6
fastembed==0.2.7
flashrank==0.2.4
groq==0.4.1
streamlit==1.30.0
nltk==3.8.1
python-dotenv==1.0.0